Fix settings
