+++ CODEOWNERS
@@ -1,1 +1,1 @@
-* @psiemens @Kay-Zee @janezpodhostnik @sideninja @DylanTinianov
+* @psiemens @Kay-Zee @janezpodhostnik @sideninja
